pyproject.toml
[project]
name = "cli-ai-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.21.1",
    "python-dotenv>=1.1.0",
]
